Dessert clicker Update
Project update

- Update MainActivity and remove a couple of
function and composable to DessertClicker file
- Update libs.version.toml with viewmodel
- Clean up a couple of file
ui/theme/Color.kt
ui/theme/Theme.kt
